Separate app from server for testing prep
