+++ app.js
@@ -1,0 +1,129 @@
+const fs = require('fs')
+const path = require('path')
+const express = require('express')
+const config = require('./lib/config.js')
+const packageJson = require('./package.json')
+
+/*  Env/Cli Config stuff
+============================================================================= */
+const BASE_URL = config.get('baseUrl')
+const GOOGLE_CLIENT_ID = config.get('googleClientId')
+const GOOGLE_CLIENT_SECRET = config.get('googleClientSecret')
+const PUBLIC_URL = config.get('publicUrl')
+const DEBUG = config.get('debug')
+const PASSPHRASE = config.get('passphrase')
+
+if (DEBUG) {
+  console.log('Config Values:')
+  console.log(config.getAllValues())
+}
+
+/*  Express setup
+============================================================================= */
+const bodyParser = require('body-parser')
+const favicon = require('serve-favicon')
+const cookieParser = require('cookie-parser')
+const cookieSession = require('cookie-session')
+const morgan = require('morgan')
+const passport = require('passport')
+const errorhandler = require('errorhandler')
+
+const app = express()
+
+app.locals.title = 'SQLPad'
+app.locals.version = packageJson.version
+app.set('env', DEBUG ? 'development' : 'production')
+
+if (DEBUG) {
+  app.use(errorhandler())
+}
+app.use(favicon(path.join(__dirname, '/public/images/favicon.ico')))
+app.use(bodyParser.json())
+app.use(
+  bodyParser.urlencoded({
+    extended: true
+  })
+)
+
+app.use(cookieParser(PASSPHRASE))
+app.use(cookieSession({ secret: PASSPHRASE }))
+app.use(passport.initialize())
+app.use(passport.session())
+app.use(BASE_URL, express.static(path.join(__dirname, 'build')))
+if (DEBUG) {
+  app.use(morgan('dev'))
+}
+app.use(function(req, res, next) {
+  // Bootstrap res.locals with any common variables
+  res.locals.message = null
+  res.locals.navbarConnections = []
+  res.locals.debug = null
+  res.locals.query = null
+  res.locals.queryMenu = false
+  res.locals.session = req.session || null
+  res.locals.pageTitle = ''
+  res.locals.user = req.user
+  res.locals.isAuthenticated = req.isAuthenticated()
+  res.locals.baseUrl = BASE_URL
+  next()
+})
+
+/*  Passport setup
+============================================================================= */
+require('./middleware/passport.js')
+
+/*  Routes
+============================================================================= */
+const routers = [
+  require('./routes/homepage.js'),
+  require('./routes/app.js'),
+  require('./routes/version.js'),
+  require('./routes/users.js'),
+  require('./routes/forgot-password.js'),
+  require('./routes/password-reset.js'),
+  require('./routes/connections.js'),
+  require('./routes/queries.js'),
+  require('./routes/query-result.js'),
+  require('./routes/download-results.js'), // streams result download to browser
+  require('./routes/schema-info.js'),
+  require('./routes/config-values.js'),
+  require('./routes/tags.js'),
+  require('./routes/signup-signin-signout.js')
+]
+
+if (GOOGLE_CLIENT_ID && GOOGLE_CLIENT_SECRET && PUBLIC_URL) {
+  if (DEBUG) {
+    console.log('Enabling Google authentication Strategy.')
+  }
+  routers.push(require('./routes/oauth.js'))
+}
+
+routers.forEach(function(router) {
+  app.use(BASE_URL, router)
+})
+
+// For any missing api route, return a 404
+app.use(BASE_URL + '/api/', function(req, res) {
+  console.log('reached catch all api route')
+  res.sendStatus(404)
+})
+
+// Anything else should render the client-side app
+// Client-side routing will take care of things from here
+// index-template.html generated by create-react-app must consider BASE_URL
+const htmlPath = path.join(__dirname, '/build/index-template.html')
+if (fs.existsSync(htmlPath)) {
+  const html = fs.readFileSync(htmlPath, 'utf8')
+  const baseUrlHtml = html
+    .replace(/="\/stylesheets/g, '="' + BASE_URL + '/stylesheets')
+    .replace(/="\/javascripts/g, '="' + BASE_URL + '/javascripts')
+    .replace(/="\/images/g, '="' + BASE_URL + '/images')
+    .replace(/="\/fonts/g, '="' + BASE_URL + '/fonts')
+    .replace(/="\/static/g, '="' + BASE_URL + '/static')
+  app.use((req, res) => res.send(baseUrlHtml))
+} else {
+  console.error('\nNO FRONT END TEMPLATE DETECTED')
+  console.error('If not running in dev mode please report this issue.\n')
+}
+
+module.exports = app
